added checks and warnings. moved example logic into its own file
